[MODIFY] splash 컴포넌트 삭제 + 리베이스
